perf: Separate label price countup from label component to reduce rerender
